refactor($core): using camel cases at file name
