+++ packages/@vuepress/core/lib/internal-plugins/enhanceApp.js
