+++ packages/@vuepress/core/lib/internal-plugins/importAsyncComponent.js
