+++ packages/@vuepress/core/lib/internal-plugins/rootMixins.js
